Fix reading band interleaved files that are greater than 2^32 bytes in size.
git-svn-id: https://svn.osgeo.org/gdal@19270 f0d54148-0727-0410-94bb-9a71ac55c965

+++ sandbox/warmerdam/pcidsk/src/core/cpcidskfile.cpp
@@ -446,8 +446,8 @@ void CPCIDSKFile::InitializeFromHeader()
                                                    image_offset, pixel_type );
 
             
-            image_offset += DataTypeSize(channel->GetType())
-                * width * height;
+            image_offset += (int64)DataTypeSize(channel->GetType())
+                * (int64)width * (int64)height;
         }
 
         else if( interleaving == "PIXEL" )
